refactor tab building logic
