+++ app/lib/pages/diary/diary_tab_widget.dart
@@ -1,16 +1,19 @@
+import 'dart:convert';
+
 import 'package:flutter/material.dart';
 import 'package:friend_private/backend/schema/memory_connection.dart';
 import 'package:friend_private/generated/l10n.dart';
 import 'package:friend_private/pages/diary/memory_chains_view.dart';
+import 'package:friend_private/backend/schema/diary.dart';
 
 class DiaryTabView extends StatelessWidget {
   final List<MemoryConnectionNode> forest;
-  final Widget diaryContent;
+  final ServerDiary? currentDateDiary;
 
   const DiaryTabView({
     Key? key,
     required this.forest,
-    required this.diaryContent,
+    required this.currentDateDiary,
   }) : super(key: key);
 
   @override
@@ -24,13 +27,13 @@ class DiaryTabView extends StatelessWidget {
               Tab(
                 child: Text(
                   S.current.MemoryChains,
-                  style: TextStyle(fontSize: 20),
+                  style: TextStyle(fontSize: 16),
                 ),
               ),
               Tab(
                 child: Text(
                   S.current.MemoryContents,
-                  style: TextStyle(fontSize: 20),
+                  style: TextStyle(fontSize: 16),
                 ),
               ),
             ],
@@ -38,16 +41,48 @@ class DiaryTabView extends StatelessWidget {
           Expanded(
             child: TabBarView(
               children: [
+                MemoryChainsView(forest: forest),
                 SingleChildScrollView(
-                  child: MemoryChainsView(forest: forest),
-                  physics: const BouncingScrollPhysics(),
-                ),
-                SingleChildScrollView(
-                  child: diaryContent,
-                  physics: const BouncingScrollPhysics(),
+                  child: _buildDiaryContent(),
                 ),
               ],
             ),
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildDiaryContent() {
+    if (currentDateDiary == null) {
+      return Padding(
+        padding: const EdgeInsets.all(16.0),
+        child: Text(
+          S.current.NoDiaryNote,
+          textAlign: TextAlign.center,
+          style: TextStyle(fontSize: 18),
+        ),
+      );
+    }
+
+    return Padding(
+      padding: const EdgeInsets.all(16.0),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          // if (diary.content.footprintJpeg == null ||
+          //     diary.content.footprintJpeg!.isEmpty)
+          //   Text(
+          //       'Footprint image not available, try set app location permission to always allow to enable this feature',
+          //       style: TextStyle(color: const Color.fromARGB(255, 189, 0, 157)))
+          // else
+          //   Image.memory(
+          //     base64Decode(diary.content.footprintJpeg!),
+          //   ),
+          // SizedBox(height: 16),
+          Text(
+            utf8.decode(currentDateDiary!.content.content.codeUnits),
+            style: TextStyle(fontSize: 16),
           ),
         ],
       ),
